[HUMAN App]: refactor: simplify operator profile page (#3158)
