+++ packages/apps/human-app/frontend/src/modules/homepage/components/operator-sign-in.tsx
@@ -4,11 +4,11 @@ import { Link } from 'react-router-dom';
 import Snackbar from '@mui/material/Snackbar';
 import { Button } from '@/shared/components/ui/button';
 import { useWalletConnect } from '@/shared/contexts/wallet-connect';
-import { useWeb3SignIn } from '@/modules/operator/hooks/use-web3-signin';
 import { useWeb3Auth } from '@/modules/auth-web3/hooks/use-web3-auth';
 import { routerPaths } from '@/router/router-paths';
 import { getErrorMessageForError } from '@/shared/errors';
 import { PrepareSignatureType } from '@/api/hooks/use-prepare-signature';
+import { useWeb3SignIn } from '../hooks';
 
 export function OperatorSignIn() {
   const { isConnected, openModal, address } = useWalletConnect();
