+++ packages/apps/human-app/frontend/src/modules/operator/hooks/index.ts
@@ -1,3 +1,1 @@
-export * from './use-disable-operator';
 export * from './use-get-keys';
-export * from './use-web3-signin';
