+++ packages/apps/human-app/frontend/src/modules/operator/profile/components/index.ts
@@ -1,2 +1,4 @@
 export * from './profile-disable-button';
 export * from './profile-enable-button';
+export * from './operator-info';
+export * from './operator-stats';
