+++ packages/apps/human-app/frontend/src/modules/operator/profile/components/profile-disable-button.tsx
@@ -1,6 +1,5 @@
 import { t } from 'i18next';
 import { Typography } from '@mui/material';
-import { useDisableWeb3Operator } from '@/modules/operator/hooks/use-disable-operator';
 import { useConnectedWallet } from '@/shared/contexts/wallet-connect';
 import { Button } from '@/shared/components/ui/button';
 import type { SignatureData } from '@/api/hooks/use-prepare-signature';
@@ -8,6 +7,7 @@ import {
   PrepareSignatureType,
   usePrepareSignature,
 } from '@/api/hooks/use-prepare-signature';
+import { useDisableWeb3Operator } from '../hooks';
 
 export function ProfileDisableButton() {
   const { address, signMessage } = useConnectedWallet();
