+++ packages/apps/human-app/frontend/src/modules/operator/profile/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './use-get-stats';
+export * from './use-disable-operator';
