+++ packages/apps/human-app/frontend/src/modules/operator/profile/hooks/use-disable-operator.ts
